Site-species matrix of plants along a Tropical Gradient

Presences and absences of plants at different elevations along a tropical gradient. Species names are indicated as column names.

site_sp_plants

Format

A matrix with 392 columns (1 per species) and 10 rows:

elev_250

is species present at 250 m elevation? 0=No, 1=Yes

elev_500

is species present at 500 m elevation? 0=No, 1=Yes

elev_1000

is species present at 1000 m elevation? 0=No, 1=Yes

elev_1500

is species present at 1500 m elevation? 0=No, 1=Yes

elev_2000

is species present at 2000 m elevation? 0=No, 1=Yes

elev_2500

is species present at 2500 m elevation? 0=No, 1=Yes

elev_3000

is species present at 3000 m elevation? 0=No, 1=Yes

elev_3500

is species present at 3500 m elevation? 0=No, 1=Yes

elev_3750

is species present at 3750 m elevation? 0=No, 1=Yes

elev_4000

is species present at 4000 m elevation? 0=No, 1=Yes

Source

Nowak, Larissa et al. (2019), Data from: Projecting consequences of global warming for the functional diversity of fleshy-fruited plants and frugivorous birds along a tropical elevational gradient, Dryad, Dataset, doi: 10.5061/dryad.c0n737b