Functional Traits of Frugivorous Birds along a Tropical Gradient

A dataset containing some functional traits of frugivorous birds in the Manú biosphere reserve, southeast Peru. Given are species mean trait values. The row names of the dataset give species names. Morphological traits have been measured on museum specimen following Eck et al.(2011). Traits have been measured only for adult and, if possible, for a minimum of two female and two male specimens. Body mass was taken from Dunning et al. (2007).

Usage

traits_birds

Format

A data frame with 217 rows and 4 variables:

Bill.width..mm.

bill width, in mm

Bill.length..mm.

bill length, in mm

Kipp.s.index

Kipp's index indicating wing Pointedness

Bodymass..g.

adult's bodymass, in g

Source

Nowak, Larissa et al. (2019), Data from: Projecting consequences of global warming for the functional diversity of fleshy-fruited plants and frugivorous birds along a tropical elevational gradient, Dryad, Dataset, doi:10.5061/dryad.c0n737b